AB#187 Integration with API backend create project
